Revert "Reverted to integral precedence scoring"
This reverts commit a8ad020abafe17f03cd801b094d5b65aba6c32f2.

+++ Classy/Parser/CASStyleSelector.h
@@ -82,7 +82,7 @@
  *
  *  @return Precendence score
  */
-- (NSInteger)precedence;
+- (CGFloat)precedence;
 
 /**
  *  Whether is selector matches the given item
